feat: Consolidated common logic into a single module
