--- src/hooks/column/useDeleteColumn.js
@@ -1,12 +1,0 @@
-import Columns from '@/api/columns';
-
-const useDeleteColumn = ({ dashboardId, columnId }) => {
-  const handleDeleteClick = async () => {
-    await Columns.delete(columnId);
-    await Columns.getList(dashboardId);
-  };
-
-  return { handleDeleteClick };
-};
-
-export default useDeleteColumn;
